refactor: wellknownAddresses to be part of registry
Ethereum L1/L2s have some well known addresses and this seems to be the
direction more dapps seem to want to go anyway. The first way I tried to
fix this is with a global hashmap, but that's ugly. Instead now it's
part of registry so that any contract can use it. This will help make
sure the contracts like Permit2 from Uniswap are same across chains,
while allowing some overrides if needed in a typesafe way.
This also adds tests to make sure that the enums are covered